requirements.txt

streamlit
requests
pandas
altair
folium
streamlit-folium
Pillow
matplotlib
